+++ packages/atproto/lib/src/services/entities/repo.dart
@@ -17,6 +17,8 @@ class Repo with _$Repo {
     required String did,
     @JsonKey(name: 'head') required String headCid,
     required String rev,
+    @Default(true) bool active,
+    String? status,
   }) = _Repo;
 
   factory Repo.fromJson(Map<String, Object?> json) => _$RepoFromJson(json);